+++ recipes/gpgme.rb
@@ -5,7 +5,7 @@
 # Cookbook Name:: terraform
 # Recipe:: gpgme
 #
-# Copyright 2016-2017, Dang Nguyen <haidangwa@gmail.com>
+# Copyright 2016-2018, Dang Nguyen <haidangwa@gmail.com>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,29 +21,31 @@
 #
 require 'tmpdir'
 
-node.default['build-essential']['compile_time'] = true
-include_recipe 'build-essential'
+# this resource installs haveged for additional entropy
+gpg_install 'gnupg2 and haveged'
+
+hashicorp_keyfile = File.join(Dir.tmpdir, 'hashicorp.asc')
 
 # deploy the hashicorp public key onto the target node
 cookbook_file 'hashicorp.asc' do
-  path File.join(Dir.tmpdir, 'hashicorp.asc')
-  mode '644'
-  action :nothing
-end.run_action(:create)
+  path hashicorp_keyfile
+  mode '0644'
+  notifies :import, 'gpg_key[import HashiCorp key to root keychain]',
+           :immediately
+end
+
+gpg_key 'import HashiCorp key to root keychain' do
+  user 'root'
+  key_file hashicorp_keyfile
+  name_real 'HashiCorp Security'
+  action :import
+end
 
 # download the signature file and the raw checksums from Hashicorp
 [sigfile, checksums_file].each do |file|
   remote_file file do
     path File.join(Dir.tmpdir, file)
-    source "#{File.dirname(terraform_url)}/#{file}"
-    mode '644'
-    action :nothing
-  end.run_action(:create)
+    source "#{terraform_url_base}/#{file}"
+    mode '0644'
+  end
 end
-
-chef_gem 'gpgme' do
-  compile_time true if respond_to?(:compile_time)
-end
-
-require 'gpgme'
-import_gpg_key