+++ test/fixtures/cookbooks/terraform_test/metadata.rb
@@ -1,0 +1,4 @@
+name 'terraform_test'
+version '0.0.1'
+description 'this cookbook is used in Test Kitchen only'
+depends 'terraform'